feature(jsnlog): angular errors batched and send to server
